Update install.sh to create src directory if not exists

+++ install.sh
@@ -2,8 +2,13 @@
 
 docker-compose down
 
-rm -fr ./src/*
-rm -frv ./src/.* 2>/dev/null
+if [[ -d ./src ]]
+then
+  rm -fr ./src/*
+  rm -frv ./src/.* 2>/dev/null
+else
+  mkdir "src"
+fi
 
 docker-compose run --rm composer create-project --prefer-dist laravel/laravel .
 docker-compose up -d --build nginx php mysql
